- Move to separate file that does not depend on grpc - annotate with @MustBeClosed
