+++ core/src/test/java/io/grpc/ContextsTest.java
@@ -28,21 +28,13 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-import com.google.common.util.concurrent.Futures;
-import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.MoreExecutors;
-import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.internal.FakeClock;
 import io.grpc.internal.NoopServerCall;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -54,8 +46,7 @@ import org.junit.runners.JUnit4;
 public class ContextsTest {
   private static Context.Key<Object> contextKey = Context.key("key");
   /** For use in comparing context by reference. */
-  private Object contextValue = new Object();
-  private Context uniqueContext = Context.ROOT.withValue(contextKey, contextValue);
+  private Context uniqueContext = Context.ROOT.withValue(contextKey, new Object());
   @SuppressWarnings("unchecked")
   private ServerCall<Object, Object> call = new NoopServerCall<Object, Object>();
   private Metadata headers = new Metadata();
@@ -270,257 +261,4 @@ public class ContextsTest {
       assertEquals("context must not be null", npe.getMessage());
     }
   }
-
-  @Test
-  public void cancellableCtxApplier_runnable_enters() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    assertSame(Context.ROOT, Context.current());
-    Contexts.cancellableContextApplier(cancellableContext).run(new Runnable() {
-      @Override
-      public void run() {
-        verifyInCancellableContext();
-      }
-    });
-    assertSame(Context.ROOT, Context.current());
-  }
-
-  @Test
-  public void cancellableCtxApplier_callable_enters() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    assertSame(Context.ROOT, Context.current());
-    Contexts.cancellableContextApplier(cancellableContext).call(new Callable<Void>() {
-      @Override
-      public Void call() throws Exception {
-        verifyInCancellableContext();
-        return null;
-      }
-    });
-    assertSame(Context.ROOT, Context.current());
-  }
-
-  @Test
-  public void cancellableCtxApplier_asyncCallable_enters() {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    assertSame(Context.ROOT, Context.current());
-    @SuppressWarnings("unused")
-    ListenableFuture<Integer> ignored = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            verifyInCancellableContext();
-            return Futures.immediateFuture(1);
-          }
-        });
-    assertSame(Context.ROOT, Context.current());
-  }
-
-  @Test
-  public void cancellableCtxApplier_runnable() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    Contexts.cancellableContextApplier(cancellableContext).run(new Runnable() {
-      @Override
-      public void run() {
-      }
-    });
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void cancellableCtxApplier_callable() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    Contexts.cancellableContextApplier(cancellableContext).call(new Callable<Void>() {
-      @Override
-      public Void call() throws Exception {
-        return null;
-      }
-    });
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void cancellableCtxApplier_asyncCallable_already_success() {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    @SuppressWarnings("unused")
-    ListenableFuture<Integer> ignored = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return Futures.immediateFuture(1);
-          }
-        });
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void cancellableCtxApplier_asyncCallable_already_fail() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    final FakeException fakeException = new FakeException();
-    ListenableFuture<Integer> future = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return Futures.immediateFailedFuture(fakeException);
-          }
-        });
-    assertTrue(cancellableContext.isCancelled());
-    boolean exceptionCaught = false;
-    try {
-      future.get();
-    } catch (ExecutionException e) {
-      if (e.getCause() == fakeException) {
-        exceptionCaught = true;
-      }
-    }
-    assertTrue(exceptionCaught);
-  }
-
-  @Test
-  public void applier_asyncCallable_success() {
-    final SettableFuture<Integer> settable = SettableFuture.create();
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    ListenableFuture<Integer> ret = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return settable;
-          }
-        });
-    assertFalse(ret.isDone());
-    assertFalse(cancellableContext.isCancelled());
-    settable.set(1);
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void applier_asyncCallable_fail() {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    final SettableFuture<Integer> settable = SettableFuture.create();
-    @SuppressWarnings("unused")
-    ListenableFuture<Integer> ignored = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return settable;
-          }
-        });
-    assertFalse(cancellableContext.isCancelled());
-    settable.setException(new FakeException());
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void applier_asyncCallable_userCancel() {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    ListenableFuture<Integer> ret = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return SettableFuture.create();
-          }
-        });
-    assertFalse(ret.isCancelled());
-    assertFalse(cancellableContext.isCancelled());
-    ret.cancel(true);
-    assertTrue(cancellableContext.isCancelled());
-  }
-
-  @Test
-  public void applier_asyncCallable_alreadyCancelledWhenListenersNotified() {
-    final Context.CancellableContext cancellableContext = makeCancellableContext();
-    ListenableFuture<Integer> ret = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            return SettableFuture.create();
-          }
-        });
-    assertFalse(ret.isCancelled());
-    assertFalse(cancellableContext.isCancelled());
-
-    final AtomicBoolean listenerFired = new AtomicBoolean();
-    ret.addListener(new Runnable() {
-      @Override
-      public void run() {
-        assertTrue(cancellableContext.isCancelled());
-        listenerFired.set(true);
-      }
-    }, MoreExecutors.directExecutor());
-    assertFalse(listenerFired.get());
-    ret.cancel(true);
-    assertTrue(listenerFired.get());
-  }
-
-  @Test(expected = FakeException.class)
-  public void applier_runnable_throws() {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    try {
-      Contexts.cancellableContextApplier(cancellableContext).run(
-          new Runnable() {
-            @Override
-            public void run() {
-              throw new FakeException();
-            }
-          });
-    } finally {
-      assertTrue(cancellableContext.isCancelled());
-    }
-  }
-
-  @Test(expected = FakeException.class)
-  public void applier_callable_throws() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    try {
-      Contexts.cancellableContextApplier(cancellableContext).call(
-          new Callable<Void>() {
-            @Override
-            public Void call() throws Exception {
-              throw new FakeException();
-            }
-          }
-      );
-    } finally {
-      assertTrue(cancellableContext.isCancelled());
-    }
-  }
-
-  @Test
-  public void applier_asyncCallable_throws() throws Exception {
-    Context.CancellableContext cancellableContext = makeCancellableContext();
-    final FakeException fakeException = new FakeException();
-    ListenableFuture<Integer> future = Contexts.cancellableContextApplier(cancellableContext)
-        .callAsync(new Callable<ListenableFuture<Integer>>() {
-          @Override
-          public ListenableFuture<Integer> call() throws Exception {
-            throw fakeException;
-          }
-        });
-    assertTrue(cancellableContext.isCancelled());
-    boolean exceptionCaught = false;
-    try {
-      future.get();
-      // should not be reached
-      Assert.fail();
-    } catch (ExecutionException expected) {
-      if (expected.getCause() == fakeException) {
-        exceptionCaught = true;
-      }
-    }
-    assertTrue(exceptionCaught);
-  }
-
-  private Context.CancellableContext makeCancellableContext() {
-    return Context.current().withValue(contextKey, contextValue).withCancellation();
-  }
-
-  private void verifyInCancellableContext() {
-    // Context.current() always returns an uncancellable surrogate, so it doesn't make sense to
-    // check that the current context is a CancellableContext. So let's just make sure it looks
-    // equivalent to the context created by makeCancelllableContext.
-    assertSame(contextValue, contextKey.get());
-  }
-
-  /**
-   * A class that lets us have a bit of type safety when testing thrown exceptions.
-   */
-  private static class FakeException extends RuntimeException { }
 }
